Cleanup and migrate naming from image to attachment
